sdr_multi_agent/mcp_servers/research_server/requirements.txt
mcp>=1.0.0
asyncio-mqtt>=0.16.0
typing-extensions>=4.0.0
google-search-results>=2.4.2
firecrawl-py>=0.0.6
python-dotenv>=1.0.0 